+++ packages/ubcli/CHANGELOG.md
@@ -6,7 +6,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-
+  - `ubcli execSql`: `-withResult` option added - if passed then expect last statement in the batch to be a statement what
+     returns a result, exec it using runSQL and returns a result as JSON
+     
+  ```shell script
+   ubcli execSql -sql 'select * from uba_user' -withResult
+   # run a statement and output colored beautified result
+   ubcli execSql -sql 'select * from uba_user' -withResult -noLogo | sed -n "/--BEGIN/,/--END/p" | tail -n +2 | head -n -2 | jq -r .
+  ```   
+      
 ### Changed
 
 ### Deprecated